added shared module
