--- kernel/src/alloc/mod.rs
@@ -1,2 +1,0 @@
-pub mod heap;
-pub mod vec;
