+++ kernel/src/exceptions.rs
@@ -1,7 +1,6 @@
-use libfenix::gpio::{
-    self,
-    pins::{GPIO1_21, GPIO1_22, GPIO1_23},
-};
+use shared::gpio::pins::{GPIO1_21, GPIO1_22, GPIO1_23};
+
+use super::gpio;
 
 #[no_mangle]
 fn data_abort_handler() {
@@ -21,5 +20,10 @@ fn fetch_abort_handler() {
 fn undefined_handler() {
     gpio::write(GPIO1_22, true);
 
+    loop {}
+}
+
+#[panic_handler]
+fn panic(_info: &core::panic::PanicInfo) -> ! {
     loop {}
 }
