+++ shared/src/alloc/mod.rs
@@ -1,0 +1,5 @@
+pub mod heap;
+
+pub extern crate alloc;
+
+pub use alloc::*;
