+++ shared/src/kernel.rs
