+++ shared/src/lib.rs
@@ -1,0 +1,8 @@
+#![no_std]
+
+pub mod alloc;
+pub mod gpio;
+pub mod interrupts;
+pub mod kernel;
+pub mod sync;
+pub mod sys;
